# packages/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SYNAPSE API Gateway",
    "version": "1.0.0-s2",
    "description": "Atlas Console BFF surface. Hand-curated S2 snapshot covering /auth, /api/v1/{decisions,orders,agents,rum,stream}. The CI gate `tests/contract/test_openapi_byte_equality.py` keeps this in lockstep with the live FastAPI app (B5)."
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Local dev (via reverse-proxy nginx on port 80)"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "operationId": "login",
        "summary": "Verify credentials and start a BFF cookie session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } }
            },
            "headers": {
              "Set-Cookie": {
                "description": "Opaque session id (HttpOnly, Secure, SameSite=Strict)",
                "schema": { "type": "string" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["auth"],
        "operationId": "logout",
        "summary": "Invalidate the current session",
        "responses": {
          "200": {
            "description": "Logged out",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StatusOk" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        },
        "security": [{ "sessionCookie": [] }]
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "operationId": "refresh",
        "summary": "Rolling refresh of the session and CSRF token",
        "responses": {
          "200": {
            "description": "Refreshed",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        },
        "security": [{ "sessionCookie": [] }]
      }
    },
...
    "/auth/reauth": {
      "post": {
        "tags": ["auth"],
        "operationId": "reauth",
        "summary": "Short-lived elevation for PII reveal (Audit Vault)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ReauthRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Elevated",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ReauthResponse" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        },
        "security": [{ "sessionCookie": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "synapse_session",
        "description": "Opaque session id minted by /auth/login. HttpOnly + Secure + SameSite=Strict."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF",
        "description": "Double-submit token returned at login. Required on POST/PUT/PATCH/DELETE."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing / expired session, or CSRF check failed",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPError" } } }
      },
...
    },
    "schemas": {
      "HTTPError": {
        "type": "object",
        "required": ["detail"],
        "properties": { "detail": { "type": "string" } }
      },
      "StatusOk": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "example": "ok" },
          "service": { "type": "string" }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 1, "maxLength": 128 },
          "password": { "type": "string", "minLength": 1, "maxLength": 256 }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["persona", "csrf_token", "expires_at", "claims"],
        "properties": {
          "persona": { "type": "string" },
          "csrf_token": { "type": "string", "description": "Send back as X-CSRF on mutating requests" },
          "expires_at": { "type": "string", "format": "date-time" },
          "claims": { "type": "object", "additionalProperties": true }
        }
      },
...
      "ReauthRequest": {
        "type": "object",
        "required": ["password"],
        "properties": { "password": { "type": "string", "minLength": 1, "maxLength": 256 } }
      },
      "ReauthResponse": {
        "type": "object",
        "required": ["status", "elevated_until"],
        "properties": {
          "status": { "type": "string", "example": "ok" },
          "elevated_until": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": []
}